backend/requirements.txt
# Core Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.15

# JWT Authentication
djangorestframework-simplejwt>=5.3.1

# CORS handling
django-cors-headers>=4.3

# Environment variables
python-dotenv>=1.0

# Database (PostgreSQL)
psycopg2-binary>=2.9

# Production server
gunicorn>=21.2

pdf2image
google-genai
pandas
openpyxl
PyMuPDF